num: Adapt tests to pddl_num_t replacing pddl_num_val_t
pddl_num_val_t and the num_val module are gone from the library, so the
num_val test module is removed too -- the num module already covers the
same ground for pddl_num_t -- and every remaining test is switched to the
pddl_num_t API (pddlNumVal*() -> pddlNum*(), pddlNumValEq() ->
pddlNumExactEq(), PDDL_OPEN_LIST_KEY_NUM_VAL -> PDDL_OPEN_LIST_KEY_NUM,
has_num_val_g_value -> has_num_g_value).

Two new tests cover what the change adds.  num_representable exercises
pddlNumIsInt64Representable() and pddlNumIsDblRepresentable(), the
predicates deciding whether a value still fits into the int/float storage
of pddl_num_t, including the boundary and the non-normal cases (INT_MAX
plus one, DBL_MIN, +-FLT_MAX, infinity, NaN).  num_err drives each of
pddlNum{Add,Sub,Mul,Div}{,To}Err() through one successful and one failing
operation per numeric status (division by zero, overflow,
delta-overflow, a huge operand, a non-zero delta, an undefined result),
checks that a failure leaves the destination untouched, and prints the
produced messages so the wording itself is part of the baseline.

The numeric-evaluation tests follow the merge of the DIV_BY_ZERO and
OVERFLOW statuses into PDDL_FM_NUM_EVAL_NUM_ERR: they now pass an error
structure to pddlFmNumExpEval()/pddlFmNumCmpEval() and check the reason
by matching a substring of the message rather than by comparing the
status alone.

The regression baselines change because floats are stored as float
instead of double, which shifts the last digits of the numeric states.
The narrower integer range makes several ztalloc-sum lifted-search tests
overflow, so they are disabled in config.toml and their baselines
removed.

+++ src/fm.c
@@ -635,7 +635,7 @@ TEST_ONCE(fm_simplify_status)
 
     printf("integer overflow:\n");
     st = runDel(pddlFmSimplify,
-                &pddlFmNewNumExpMult(pddlFmNewNumExpNumInt(INT64_MAX),
+                &pddlFmNewNumExpMult(pddlFmNewNumExpNumInt(INT_MAX),
                                      pddlFmNewNumExpNumInt(2))->fm,
                 &pddl, NULL, &err);
     assert(st == -1);
